patch correctif images

+++ backend/controllers/artists.js
@@ -224,7 +224,7 @@ export const updateArtist = async (req, res, silent = false) => {
       }
 
       // === Cas 2 : nouvelle image (remplace une image ou un logo précédent) ===
-      if (mediaType === "image") {
+     if (mediaType === "image" && sentNewMedia) {
         // Supprimer ancienne image si différente
         if (oldImageId && oldImageId !== newImageId) {
           await imagekit.deleteFile(oldImageId);
